Create generic docker for gateway
