+++ GatewayRunner/pom.xml
@@ -1,5 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <parent>
     <artifactId>scalecube-gateway-parent</artifactId>
     <groupId>io.scalecube</groupId>
@@ -7,16 +9,26 @@
   </parent>
   <modelVersion>4.0.0</modelVersion>
 
-  <artifactId>scalecube-gateway-rsocket-websocket-runner</artifactId>
+  <artifactId>scalecube-gateway-runner</artifactId>
 
   <properties>
-    <mainClass>io.scalecube.gateway.rsocket.websocket.RSocketWebsocketGatewayRunner</mainClass>
+    <mainClass>io.scalecube.gateway.GatewayRunner</mainClass>
   </properties>
 
   <dependencies>
     <dependency>
       <groupId>io.scalecube</groupId>
+      <artifactId>scalecube-gateway-http</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>io.scalecube</groupId>
       <artifactId>scalecube-gateway-rsocket-websocket</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>io.scalecube</groupId>
+      <artifactId>scalecube-gateway-websocket</artifactId>
       <version>${project.version}</version>
     </dependency>
     <dependency>
