+++ Http/src/main/java/io/scalecube/gateway/http/HttpGateway.java
@@ -5,6 +5,8 @@ import io.scalecube.services.gateway.Gateway;
 import io.scalecube.services.gateway.GatewayConfig;
 import io.scalecube.services.metrics.Metrics;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import reactor.core.publisher.Mono;
 import reactor.ipc.netty.http.server.HttpServer;
 import reactor.ipc.netty.tcp.BlockingNettyContext;
@@ -16,6 +18,8 @@ import java.util.concurrent.ExecutorService;
 
 public class HttpGateway implements Gateway {
 
+  private static final Logger LOGGER = LoggerFactory.getLogger(HttpGateway.class);
+
   private BlockingNettyContext server;
 
   @Override
@@ -23,6 +27,8 @@ public class HttpGateway implements Gateway {
       Metrics metrics) {
 
     return Mono.defer(() -> {
+      LOGGER.info("Starting gateway with {}", config);
+
       InetSocketAddress listenAddress = new InetSocketAddress(config.port());
       GatewayHttpAcceptor acceptor = new GatewayHttpAcceptor(call.create());
 
@@ -39,6 +45,9 @@ public class HttpGateway implements Gateway {
           .start(acceptor);
       server.installShutdownHook();
       InetSocketAddress address = server.getContext().address();
+
+      LOGGER.info("Gateway has been started successfully on {}", address);
+
       return Mono.just(address);
     });
   }
