--- WebsocketRunner/src/main/resources/config/config.properties
@@ -1,2 +1,0 @@
-io.scalecube.gateway.websocket.seedAddress=localhost:4801
-io.scalecube.gateway.websocket.websocketPort=9090
